Make hans argument number a float

+++ include/hans/common/types.hpp
@@ -46,7 +46,7 @@ typedef struct {
   hans_hash name;
   hans_arg_type type;
   union {
-    int32_t number;
+    float number;
     bool boolean;
     hans_hash string;
   };
